# docs/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.0"
  title: "Authentication Sytem API"
basePath: "http://localhost:3030" 

paths:
  /session/refresh:
    patch:
      summary: "Refresh Access Token"
      description: "Refreshes the access token using the provided refresh token."
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: refreshToken
          in: body
          description: "Refresh token to obtain a new access token."
          required: true
          schema:
            $ref: "#/definitions/RefreshTokenDto"
      responses:
        200:
          description: "Successful token refresh."
          schema:
            $ref: "#/definitions/AccessTokenResponse"
        404:
          description: "Invalid refresh token."

definitions:
  RefreshTokenDto:
    type: object
    properties:
      refreshToken:
        type: string
        example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ0b2tlblR5cGUiOiJyZWZyZXNoIiwiaWF0IjoxNjg5ODgwNTQyLCJleHAiOjE2ODk4ODA2MDJ9.pvb_akzcwmjSrz_DPsb_cLu0uUqHD6RIVEy_beau9TI"

  AccessTokenResponse:
    type: object
    properties:
      accessToken:
        type: string
        example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIwNjNkMmMwNy03ZTdkLTRlYTUtOWJmYi03MWY3YTU4Y2MwNjkiLCJlbWFpbCI6InJpY2FyZG8uZGlhc0BnbWFpbC5jb20iLCJpYXQiOjE2ODk4OTU4NDUsImV4cCI6MTY4OTg5NTkwNX0.dsHI_4aGu8pLcBk978gkYD99WzcKQgkdmE4ziD2SRPE"


paths:
  /user/create:
    post:
      tags:
        - user
      summary: Criar um novo usuário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterUserDto'
      responses:
        '201':
          description: Usuário criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /user/list-one/{id}:
    get:
      tags:
        - user
      summary: Buscar um usuário pelo ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Usuário encontrado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: Usuário não encontrado
  /user/update-password/{id}:
    patch:
      tags:
        - user
      summary: Atualizar a senha de um usuário
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePasswordDTO'
      responses:
        '204':
          description: Senha do usuário atualizada com sucesso
        '404':
          description: Usuário não encontrado
  /user/delete/{id}:
    delete:
      tags:
        - user
      summary: Excluir um usuário
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Usuário excluído com sucesso
        '404':
          description: Usuário não encontrado
components:
  schemas:
    RegisterUserDto:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
      required:
        - name
        - email
        - password
    UpdatePasswordDTO:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
        confirmNewPassword:
          type: string
      required:
        - currentPassword
        - newPassword
        - confirmNewPassword
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        password_hash:
          type: string
        created_at:
          type: string
        deleted:
          type: boolean